Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "azure-identity>=1.25.1",
    "azure-keyvault-secrets>=4.10.0",
    "azure-monitor-opentelemetry>=1.8.5",
    "fastapi>=0.128.0",
    "gunicorn>=24.1.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
